requirements.txt
Django>=4.1,<=4.2
djangorestframework>=3.13.1,<=3.14.0
psycopg2>=2.9.5,<=2.9.6
drf-spectacular>=0.26.1,<=0.26.2
pillow>=9.4.0,<=9.5.0
django-filter>=23.1,<=23.2
djangorestframework-gis>=0.18,<=1.0
django-cors-headers>=3.14.0,<=4.0.0
django-geo>=0.7,<=0.8